requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.25.9
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
fsspec==2024.6.1
greenlet==3.0.3
groq==0.6.0
google-generativeai==0.8.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
jinja2==3.1.4
jiter==0.5.0
jsonref==1.1.0
mako==1.3.5
markupsafe==2.1.5
openai==1.40.6
packaging==24.1
prompt-toolkit==3.0.47
psutil==5.9.8
pydantic-core==2.20.1
pydantic==2.8.2
python-dotenv==1.0.1
pyyaml==6.0.2
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
sqlalchemy==2.0.32
sqlalchemy[asyncio]==2.0.32
tenacity==9.0.0
tiktoken==0.6.0
tokenizers==0.20.0
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13

prompt_toolkit~=3.0.47
protobuf~=5.28.3
typing_extensions~=4.12.2